+++ src/easydel/inference/serve_engine/__init__.py
@@ -1,0 +1,4 @@
+from easydel.inference.serve_engine.serve import ApiEngine as ApiEngine
+from easydel.inference.serve_engine import client as engine_client
+
+engine_client = engine_client  # Ide will find this one here